Reduce emoji prism for now.
I'm pretty sure the larger prism is technically better, but the size of
the context window is currently small enough that I don't think we are
making as good of use as we could. Reducing the prism size gets us just
under 1500 characters, which allows RIP to only consume a single custom
instruction block almost perfectly.