Back to basics!
The trick is not to scare Claude in header, then prime attention. EMBODY
was one of the first keywords I tried and it conveys the implicit nature
of framework perfectly. I only moved away from it trying to get explicit
symbolic planning working, and turns out just slowing the AI down some
and adding some terms in the math related to reflection was enough!

This should also restore some of the capabilities from the README as
well, until I get a chance to fix that up properly. Need to figure out
how to organize all the different frameworks that are possible, or
create a generator!